requirements.txt
streamlit>=1.28.0
pydantic-ai==0.1.3
pydantic-ai-slim[mcp]==0.1.3
pydantic-ai-slim[groq]==0.1.3
mcp-server-fetch==2025.4.7
python-dotenv>=1.0.0